+++ src/dataflow/src/source/kafka.rs
@@ -13,19 +13,19 @@ use std::convert::TryInto;
 use std::sync::{Arc, Mutex};
 use std::time::Duration;
 
-use dataflow_types::{
-    Consistency, DataEncoding, ExternalSourceConnector, KafkaOffset, KafkaSourceConnector, MzOffset,
-};
-use expr::{PartitionId, SourceInstanceId};
-use log::{error, info, log_enabled, warn};
 use rdkafka::consumer::base_consumer::PartitionQueue;
 use rdkafka::consumer::{BaseConsumer, Consumer, ConsumerContext};
 use rdkafka::message::BorrowedMessage;
 use rdkafka::topic_partition_list::Offset;
 use rdkafka::{ClientConfig, ClientContext, Message, Statistics, TopicPartitionList};
-
 use timely::scheduling::activate::{Activator, SyncActivator};
-use url::Url;
+
+use dataflow_types::{
+    Consistency, DataEncoding, ExternalSourceConnector, KafkaOffset, KafkaSourceConnector, MzOffset,
+};
+use expr::{PartitionId, SourceInstanceId};
+use kafka_util::KafkaAddr;
+use log::{error, info, log_enabled, warn};
 
 use crate::server::{
     TimestampDataUpdate, TimestampDataUpdates, TimestampMetadataUpdate, TimestampMetadataUpdates,
@@ -340,14 +340,14 @@ impl KafkaSourceInfo {
         kc: KafkaSourceConnector,
     ) -> KafkaSourceInfo {
         let KafkaSourceConnector {
-            url,
+            addr,
             topic,
             config_options,
             group_id_prefix,
             ..
         } = kc;
         let kafka_config =
-            create_kafka_config(&source_name, &url, group_id_prefix, &config_options);
+            create_kafka_config(&source_name, &addr, group_id_prefix, &config_options);
         let source_id = source_id.to_string();
         let consumer: BaseConsumer<GlueConsumerContext> = kafka_config
             .create_with_context(GlueConsumerContext(consumer_activator))
@@ -486,14 +486,14 @@ impl KafkaSourceInfo {
 /// Creates a Kafka config.
 fn create_kafka_config(
     name: &str,
-    url: &Url,
+    addr: &KafkaAddr,
     group_id_prefix: Option<String>,
     config_options: &HashMap<String, String>,
 ) -> ClientConfig {
     let mut kafka_config = ClientConfig::new();
 
     // Broker configuration.
-    kafka_config.set("bootstrap.servers", &url.to_string());
+    kafka_config.set("bootstrap.servers", &addr.to_string());
 
     // Opt-out of Kafka's offset management facilities. Whenever we restart,
     // we want to restart from the beginning of the topic.